db_portal/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
Jinja2==2.11.1
matplotlib==3.1.3
mplleaflet==0.0.5
numpy==1.18.1
oauth2client==3.0.0
opencv-python-headless==4.2.0.32
Pillow==7.0.0
pyfcm==1.4.7
Pyrebase==3.0.27
pytesseract==0.3.2
python-dotenv==0.12.0
requests==2.11.1
Werkzeug==1.0.0
